Finished with the final combined tree optimization
Unfortunately it requried the attempted overloading of the
AddRate2Branch function, which apparently isn't possible in
HBL. In the mean time I just use a normal and "Advanced"
version of the function, with the normal one calling the
advanced one with some dummy parameters.

Everything appears to work now, likelihoods are rarely if
ever worse than they would have been with the MGL model,
and additional omega parameters rarely decrease the
likelihood.

Unfortunately the final combined version ends with a
positive log(Likelihood), which I'm not entirely sure is
possible. Time to get a second opinion.