Parameter initialization is improved
But still not perfect. In its current incarnation it takes a site
from the first omega and gives it to the last. Unfortunately there
might not be a site in the first omega. This needs to change, I'm
just not yet sure how.

As for the final tree: The correct parameters aren't sticking
to the tree. The wrong number of omegas will show up, or there
will be omegas but no constraints on their proportions.