Fixed the final model
There was a variable naming bug in the parameter addition loops
preventing the Paux constraints from being applied. The existance
of appropriate omegas was the result of carry-overs from previous
models that I didn't think possible.

The construction of the final likelihood function is now moved
to its own function. The simple/advanced forms of addRateClass2Bra
-nch can now be eliminated.

The likelihood is so far never worse than the all-MGL likelihood,
and in most cases it is better. It is also really excellent for
the final model with multiple omegas on different branches.

The remaining challenge is refining the parameter initialization
routines, as they currently still just steal one site from the
first omega. This doesn't work if the first omega has zero sites,
and isn't fair when it does work (we're assuming additional omegas
are added to the higher end of the distribution). The solution then
is to pull sites from other omegas, but doing this is proving to be
a challenge.