Finished better proportion initialization
Essentially it starts at the most recently added Paux and determines
whether or not more than two sites have been assigned to this
omega. If it has, it takes one site. It it does not, it goes to
the next omega. This seems to work pretty well and can be justified
logically.

I also redid the parameter initialization for the final composite
model. It is now based on the MGL omega value and one site per
additional omega, with those sites being taken from the first
omega's proportion.