Branch length calculator is coming along
It currently works, but the branch lengths aren't turning out
as we would expect. They're close, but no cigar. Issues:

- Do you divide the branch length by three? By the number of omegas?
By nothing at all?

- The omegas in the final model are not being initialized to
anything close to correct.

- Moving from MGL to BSREL1 for situations where no additional
omegas are required actually reduces the total number of parameters
in the model, thereby improving the BIC. Should this be used from
the outset? Should this not be used at all? It was implemented for
the branch length calculator.