Finished up initialization of final model
As it stands the final model still only has one omega for all each
MGL tree, which may be skewing results. These omegas should be more
accurately initialized, at least. The branch length results aren't
phenomenal: they are within a factor of three of expected, but some
are a factor of three too high, some are a factor of three too low.
Further experimentation will artificially set the number of branches
in the final model to see if the expected results can be teased out
by changing the number of parameters.