Discovered why things weren't working
MPI send can take a string and execute it, however objects and
variables are not present on the MPI nodes.

- The LF creation needs to be done on the root node, optimization
on the MPI nodes
- The return value needs to be properly correlated with the node
number, which needs to be properly correlated with a branch number
- Models created on the MPI nodes aren't present in the model list.
Fixing item 1 will fix most of this, except that the model list
cannot be moved off to the MPI nodes.